+++ test/integration/suite_test.go
@@ -21,8 +21,13 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
 
 	testutils "github.com/kong/kubernetes-ingress-controller/v2/internal/util/test"
+)
+
+var (
+	k8sClient *kubernetes.Clientset
 )
 
 // -----------------------------------------------------------------------------
@@ -95,6 +100,7 @@ func TestMain(m *testing.M) {
 	var err error
 	env, err = builder.Build(ctx)
 	exitOnErr(err)
+	k8sClient = env.Cluster().Client()
 
 	fmt.Printf("INFO: reconfiguring the kong admin service as LoadBalancer type\n")
 	svc, err := env.Cluster().Client().CoreV1().Services(kongAddon.Namespace()).Get(ctx, kong.DefaultAdminServiceName, metav1.GetOptions{})