+++ test/integration/utils_test.go
@@ -18,12 +18,14 @@ import (
 	"time"
 
 	"github.com/blang/semver/v4"
+	"github.com/google/uuid"
 	"github.com/kong/go-kong/kong"
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
 	"github.com/kong/kubernetes-testing-framework/pkg/environments"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // -----------------------------------------------------------------------------
@@ -461,4 +463,25 @@ func exitOnErr(err error) {
 		return
 	}
 	exitOnErrWithCode(err, ExitCodeEnvSetupFailed)
+}
+
+// TODO move this into a shared library https://github.com/Kong/kubernetes-testing-framework/issues/302
+// setup is a helper function for tests which conveniently creates a cluster
+// cleaner (to clean up test resources automatically after the test finishes)
+// and creates a new namespace for the test to use. It also enables parallel
+// testing.
+func setup(t *testing.T) (*corev1.Namespace, *clusters.Cleaner) {
+	t.Log("performing test setup")
+	cleaner := clusters.NewCleaner(env.Cluster())
+
+	t.Log("creating a testing namespace")
+	namespace, err := k8sClient.CoreV1().Namespaces().Create(ctx, &corev1.Namespace{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: uuid.NewString(),
+		},
+	}, metav1.CreateOptions{})
+	require.NoError(t, err)
+	cleaner.AddNamespace(namespace)
+
+	return namespace, cleaner
 }